Implemented BLEToolkitApp with connection and scanning capabilities
The application BLEToolkitApp now has a connection tab, which allows devices to be scanned and connected. It's using the Adafruit BLE library to perform these actions. The conncetion tab has been implemented in connection_tab.py. Also, added a new GradientButton class for styling purposes and updated the requirements.txt with necessary packages.